visual optimize add RSICurveItem
