+++ vnpy/app/realtime_monitor/ui/indicatorQtItems.py
@@ -1,0 +1,460 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# @Time    : 2019/9/3 0003 13:01
+# @Author  : Hadrianl 
+# @File    : indicatorQtItems
+
+import numpy as np
+import pyqtgraph as pg
+import datetime as dt
+
+from PyQt5 import QtGui, QtCore
+
+# DEFAULT_MA = [5, 10, 30, 60]
+
+from vnpy.chart.item import ChartItem
+from vnpy.chart.manager import BarManager
+from vnpy.trader.object import BarData, TickData, TradeData, OrderData
+from vnpy.trader.constant import Direction, Status, Interval
+from vnpy.chart.base import BAR_WIDTH, PEN_WIDTH, UP_COLOR, DOWN_COLOR, CURSOR_COLOR, BLACK_COLOR, NORMAL_FONT
+from vnpy.trader.utility import ArrayManager, BarGenerator
+from typing import Tuple, Callable
+from collections import defaultdict
+
+import talib
+
+
+class MACurveItem(ChartItem):
+    name = 'ma'
+    plot_name = 'candle'
+    MA_COLORS = {5: pg.mkPen(color=(255, 255, 255), width=PEN_WIDTH),
+                 10: pg.mkPen(color=(255, 255, 0), width=PEN_WIDTH),
+                 20: pg.mkPen(color=(218, 112, 214), width=PEN_WIDTH),
+                 30: pg.mkPen(color=(0, 255, 0), width=PEN_WIDTH),
+                 60: pg.mkPen(color=(64, 224, 208), width=PEN_WIDTH)}
+    def __init__(self, manager: BarManager):
+        """"""
+        super().__init__(manager)
+        self.periods = [5, 10, 20, 30, 60]
+        self._arrayManager = ArrayManager(max(self.periods) + 1)
+        self.mas = defaultdict(dict)
+        self.last_ix = 0
+        self.last_picture = QtGui.QPicture()
+
+    def _draw_bar_picture(self, ix: int, bar: BarData) -> QtGui.QPicture:
+        """"""
+        # Create objects
+
+        if ix <= self.last_ix:
+            return self.last_picture
+
+        pre_bar = self._manager.get_bar(ix-1)
+
+        if not pre_bar:
+            return self.last_picture
+
+        ma_picture = QtGui.QPicture()
+        self._arrayManager.update_bar(pre_bar)
+        painter = QtGui.QPainter(ma_picture)
+
+        # Draw volume body
+        for p in self.periods:
+            sma=self._arrayManager.ma(p, True)
+            pre_ma = sma[-2]
+            ma = sma[-1]
+            self.mas[p][ix-1] = ma
+            if np.isnan(pre_ma) or np.isnan(ma):
+                continue
+            line = QtCore.QLine()
+            line.setLine(ix-2, sma[-2], ix-1, sma[-1])
+            painter.setPen(self.MA_COLORS[p])
+            painter.drawLine(line)
+
+        # Finish
+        painter.end()
+        self.last_ix = ix
+        self.last_picture = ma_picture
+        return ma_picture
+
+    def boundingRect(self) -> QtCore.QRectF:
+        """"""
+        min_price, max_price = self._manager.get_price_range()
+        rect = QtCore.QRectF(
+            0,
+            min_price,
+            len(self._bar_picutures),
+            max_price - min_price
+        )
+        return rect
+
+    def get_y_range(self, min_ix: int = None, max_ix: int = None) -> Tuple[float, float]:
+        """
+        Get range of y-axis with given x-axis range.
+
+        If min_ix and max_ix not specified, then return range with whole data set.
+        """
+        min_volume, max_volume = self._manager.get_price_range(min_ix, max_ix)
+        return min_volume, max_volume
+
+    def get_info_text(self, ix: int) -> str:
+        """
+        Get information text to show by cursor.
+        """
+        text = '\n'.join(f'ma{p}: {v.get(ix, np.nan):.2f}' for p, v in self.mas.items())
+        return f"MA \n{text}"
+
+    def clear_all(self) -> None:
+        """
+        Clear all data in the item.
+        """
+        super().clear_all()
+        self._arrayManager = ArrayManager(max(self.periods) + 1)
+        self.mas = defaultdict(dict)
+        self.last_ix = 0
+        self.last_picture = QtGui.QPicture()
+
+class MACDItem(ChartItem):
+    name = 'macd'
+    plot_name = 'indicator'
+    MACD_PARAMS = [12, 26, 9]
+    MACD_COLORS = {'diff': pg.mkPen(color=(255, 255, 255), width=PEN_WIDTH),
+                 'dea': pg.mkPen(color=(255, 255, 0), width=PEN_WIDTH),
+                 'macd': {'up': pg.mkBrush(color=(255, 0, 0)), 'down': pg.mkBrush(color=(0, 255, 50))}}
+    def __init__(self, manager: BarManager):
+        """"""
+        super().__init__(manager)
+        self._arrayManager = ArrayManager(150)
+        self.last_ix = 0
+        self.last_picture = QtGui.QPicture()
+        self.macds = defaultdict(dict)
+        self.br_max = 0
+        self.br_min = 0
+
+    def _draw_bar_picture(self, ix: int, bar: BarData) -> QtGui.QPicture:
+        """"""
+        # Create objects
+        if ix <= self.last_ix:
+            return self.last_picture
+
+        pre_bar = self._manager.get_bar(ix-1)
+
+        if not pre_bar:
+            return self.last_picture
+
+        macd_picture = QtGui.QPicture()
+        self._arrayManager.update_bar(pre_bar)
+        painter = QtGui.QPainter(macd_picture)
+
+        diff, dea, macd = self._arrayManager.macd(*self.MACD_PARAMS, array=True)
+        self.br_max = max(self.br_max, diff[-1], dea[-1], macd[-1])
+        self.br_min = min(self.br_min, diff[-1], dea[-1], macd[-1])
+        self.macds['diff'][ix-1] = diff[-1]
+        self.macds['dea'][ix-1] = dea[-1]
+        self.macds['macd'][ix-1] = macd[-1]
+        if not (np.isnan(diff[-2]) or np.isnan(dea[-2]) or np.isnan(macd[-1])):
+            diff_line = QtCore.QLine()
+            dea_line = QtCore.QLine()
+            diff_line.setLine(ix - 2, diff[-2], ix - 1, diff[-1])
+            dea_line.setLine(ix - 2, dea[-2], ix - 1, dea[-1])
+            macd_bar = QtCore.QRectF(ix - 1 - BAR_WIDTH, 0,
+                                     BAR_WIDTH * 2, macd[-1])
+
+            painter.setPen(self.MACD_COLORS['dea'])
+            painter.drawLine(dea_line)
+            painter.setPen(self.MACD_COLORS['diff'])
+            painter.drawLine(diff_line)
+
+            if macd[-1] > 0:
+                painter.setBrush(self.MACD_COLORS['macd']['up'])
+            else:
+                painter.setBrush(self.MACD_COLORS['macd']['down'])
+
+            painter.drawRect(macd_bar)
+
+        # Finish
+        painter.end()
+        self.last_ix = ix
+        self.last_picture = macd_picture
+        return macd_picture
+
+    def boundingRect(self) -> QtCore.QRectF:
+        """"""
+        rect = QtCore.QRectF(
+            0,
+            self.br_min,
+            len(self._bar_picutures),
+            self.br_max - self.br_min
+        )
+        return rect
+
+
+    def get_y_range(self, min_ix: int = None, max_ix: int = None) -> Tuple[float, float]:
+        """
+        Get range of y-axis with given x-axis range.
+
+        If min_ix and max_ix not specified, then return range with whole data set.
+        """
+        min_ix = 0 if min_ix is None else min_ix
+        max_ix = self.last_ix if max_ix is None else max_ix
+
+        min_v = 0
+        max_v = 0
+
+        for i in range(min_ix, max_ix):
+            min_v = min(min_v, self.macds['diff'].get(i, min_v), self.macds['dea'].get(i, min_v), self.macds['macd'].get(i, min_v))
+            max_v = max(max_v, self.macds['diff'].get(i, max_v), self.macds['dea'].get(i, max_v), self.macds['macd'].get(i, max_v))
+
+        return min_v, max_v
+
+    def get_info_text(self, ix: int) -> str:
+        """
+        Get information text to show by cursor.
+        """
+        return f"MACD{self.MACD_PARAMS}  DIFF:{self.macds['diff'].get(ix, np.nan):.2f} DEA:{self.macds['dea'].get(ix, np.nan):.2f} MACD:{self.macds['macd'].get(ix, np.nan):.2f}"
+
+    def clear_all(self) -> None:
+        """
+        Clear all data in the item.
+        """
+        super().clear_all()
+        self._arrayManager = ArrayManager(150)
+        self.last_ix = 0
+        self.last_picture = QtGui.QPicture()
+        self.macds = defaultdict(dict)
+        self.br_max = 0
+        self.br_min = 0
+
+
+class INCItem(ChartItem):
+    name = 'inc'
+    plot_name = 'indicator'
+    INC_PARAMS = [60, 2]
+    INC_COLORS = {'up': pg.mkPen(color=(0, 0, 255), width=PEN_WIDTH),
+                 'inc': {'up_gte': pg.mkBrush(color=(255, 0, 0)), 'up_lt': pg.mkBrush(color=(160, 32, 240)),
+                         'down_gte': pg.mkBrush(color=(0, 255, 0)), 'down_lt': pg.mkBrush(color=(0, 255, 255))},
+                 'down': pg.mkPen(color=(255, 255, 0), width=PEN_WIDTH)}
+    def __init__(self, manager: BarManager):
+        """"""
+        super().__init__(manager)
+        self._arrayManager = ArrayManager(150)
+        self.last_ix = 0
+        self.last_picture = QtGui.QPicture()
+        self.incs = defaultdict(dict)
+        self.br_max = 0
+        self.br_min = 0
+
+    def _draw_bar_picture(self, ix: int, bar: BarData) -> QtGui.QPicture:
+        """"""
+        # Create objects
+        if ix <= self.last_ix:
+            return self.last_picture
+
+        pre_bar = self._manager.get_bar(ix-1)
+
+        if not pre_bar:
+            return self.last_picture
+
+        inc_picture = QtGui.QPicture()
+        self._arrayManager.update_bar(pre_bar)
+        painter = QtGui.QPainter(inc_picture)
+
+        inc = self._arrayManager.close - self._arrayManager.open
+        std = talib.STDDEV(inc, self.INC_PARAMS[0])
+        multiple =  inc / std
+
+        # diff, dea, macd = self._arrayManager.macd(*self.MACD_PARAMS, array=True)
+        self.br_max = max(self.br_max, std[-1], inc[-1])
+        self.br_min = min(self.br_min, -std[-1], inc[-1])
+        self.incs['up'][ix-1] = std[-1]
+        self.incs['inc'][ix-1] = inc[-1]
+        self.incs['down'][ix-1] = -std[-1]
+        self.incs['multiple'][ix-1] = multiple[-1]
+        if not (np.isnan(std[-2]*std[-1]*inc[-2]*inc[-1])):
+            up_line = QtCore.QLine()
+            down_line = QtCore.QLine()
+            up_line.setLine(ix - 2, std[-2], ix - 1, std[-1])
+            down_line.setLine(ix - 2, -std[-2], ix - 1, -std[-1])
+            multiple_bar = QtCore.QRectF(ix - 1 - BAR_WIDTH, 0,
+                                         BAR_WIDTH * 2, inc[-1])
+
+            painter.setPen(self.INC_COLORS['up'])
+            painter.drawLine(up_line)
+            painter.setPen(self.INC_COLORS['down'])
+            painter.drawLine(down_line)
+
+            painter.setPen(pg.mkPen(color=(255, 255, 255), width=PEN_WIDTH/2))
+
+            if multiple[-1] >= 0:
+                ud = 'up'
+            else:
+                ud = 'down'
+
+            if abs(multiple[-1]) >= self.INC_PARAMS[1]:
+                cp = 'gte'
+            else:
+                cp = 'lt'
+
+            painter.setBrush(self.INC_COLORS['inc'][f'{ud}_{cp}'])
+
+            painter.drawRect(multiple_bar)
+
+        # Finish
+        painter.end()
+        self.last_ix = ix
+        self.last_picture = inc_picture
+        return inc_picture
+
+    def boundingRect(self) -> QtCore.QRectF:
+        """"""
+        rect = QtCore.QRectF(
+            0,
+            self.br_min,
+            len(self._bar_picutures),
+            self.br_max - self.br_min
+        )
+        return rect
+
+
+    def get_y_range(self, min_ix: int = None, max_ix: int = None) -> Tuple[float, float]:
+        """
+        Get range of y-axis with given x-axis range.
+
+        If min_ix and max_ix not specified, then return range with whole data set.
+        """
+        min_ix = 0 if min_ix is None else min_ix
+        max_ix = self.last_ix if max_ix is None else max_ix
+
+        min_v = 0
+        max_v = 0
+
+        for i in range(min_ix, max_ix):
+            min_v = min(min_v, self.incs['down'].get(i, min_v), self.incs['inc'].get(i, min_v))
+            max_v = max(max_v, self.incs['up'].get(i, max_v), self.incs['inc'].get(i, max_v))
+
+        return min_v, max_v
+
+    def get_info_text(self, ix: int) -> str:
+        """
+        Get information text to show by cursor.
+        """
+        return f"INC{self.INC_PARAMS}  UP:{self.incs['up'].get(ix, np.nan):.2f} DOWN:{self.incs['down'].get(ix, np.nan):.2f} INC:{self.incs['inc'].get(ix, np.nan):.2f} MUTIPLE:{self.incs['multiple'].get(ix, np.nan):.2f}"
+
+    def clear_all(self) -> None:
+        """
+        Clear all data in the item.
+        """
+        super().clear_all()
+        self._arrayManager = ArrayManager(150)
+        self.last_ix = 0
+        self.last_picture = QtGui.QPicture()
+        self.incs = defaultdict(dict)
+        self.br_max = -np.inf
+        self.br_min = np.inf
+
+
+class RSICurveItem(ChartItem):
+    name = 'rsi'
+    plot_name = 'indicator'
+    RSI_COLORS = {6: pg.mkPen(color=(255, 255, 255), width=PEN_WIDTH),
+                 12: pg.mkPen(color=(255, 255, 0), width=PEN_WIDTH),
+                 24: pg.mkPen(color=(218, 112, 214), width=PEN_WIDTH)}
+    def __init__(self, manager: BarManager):
+        """"""
+        super().__init__(manager)
+        self.periods = [6, 12, 24]
+        self._arrayManager = ArrayManager(150)
+        self.rsis = defaultdict(dict)
+        self.last_ix = 0
+        self.br_max = -np.inf
+        self.br_min = np.inf
+        self.last_picture = QtGui.QPicture()
+
+    def _draw_bar_picture(self, ix: int, bar: BarData) -> QtGui.QPicture:
+        """"""
+        # Create objects
+
+        if ix <= self.last_ix:
+            return self.last_picture
+
+        pre_bar = self._manager.get_bar(ix-1)
+
+        if not pre_bar:
+            return self.last_picture
+
+        rsi_picture = QtGui.QPicture()
+        self._arrayManager.update_bar(pre_bar)
+        painter = QtGui.QPainter(rsi_picture)
+
+        # Draw volume body
+        for p in self.periods:
+            rsi_=self._arrayManager.rsi(p, True)
+            pre_rsi = rsi_[-2]
+            rsi = rsi_[-1]
+            self.rsis[p][ix-1] = rsi
+            if np.isnan(pre_rsi) or np.isnan(rsi):
+                continue
+
+            self.br_max = max(self.br_max, rsi_[-1])
+            self.br_min = min(self.br_min, rsi_[-1])
+
+            line = QtCore.QLine()
+            line.setLine(ix-2, rsi_[-2], ix-1, rsi_[-1])
+            painter.setPen(self.RSI_COLORS[p])
+            painter.drawLine(line)
+
+        # Finish
+        painter.end()
+        self.last_ix = ix
+        self.last_picture = rsi_picture
+        return rsi_picture
+
+    def boundingRect(self) -> QtCore.QRectF:
+        """"""
+        rect = QtCore.QRectF(
+            0,
+            self.br_max,
+            len(self._bar_picutures),
+            self.br_max - self.br_min
+        )
+        return rect
+
+    def get_y_range(self, min_ix: int = None, max_ix: int = None) -> Tuple[float, float]:
+        """
+        Get range of y-axis with given x-axis range.
+
+        If min_ix and max_ix not specified, then return range with whole data set.
+        """
+        min_ix = 0 if min_ix is None else min_ix
+        max_ix = self.last_ix if max_ix is None else max_ix
+
+        min_v = np.inf
+        max_v = -np.inf
+
+        p = self.periods[0]
+        for i in range(min_ix, max_ix):
+            min_v = min(min_v, self.rsis[p].get(i, min_v), self.rsis[p].get(i, min_v))
+            max_v = max(max_v, self.rsis[p].get(i, max_v), self.rsis[p].get(i, max_v))
+
+        return min_v, max_v
+
+    def get_info_text(self, ix: int) -> str:
+        """
+        Get information text to show by cursor.
+        """
+        text = '\n'.join(f'rsi{p}: {v.get(ix, np.nan):.2f}' for p, v in self.rsis.items())
+        return f"RSI \n{text}"
+
+    def clear_all(self) -> None:
+        """
+        Clear all data in the item.
+        """
+        super().clear_all()
+        self._arrayManager = ArrayManager(150)
+        self.last_ix = 0
+        self.last_picture = QtGui.QPicture()
+        self.rsis = defaultdict(dict)
+        self.br_max = -np.inf
+        self.br_min = np.inf
+
+
+INDICATOR = [MACurveItem, MACDItem, INCItem, RSICurveItem]
